make User optional so I dont break everything

+++ src/sentry/sentry_apps/api/endpoints/installation_external_issue_details.py
@@ -6,11 +6,9 @@ from sentry.api.base import control_silo_endpoint
 from sentry.sentry_apps.api.bases.sentryapps import (
     SentryAppInstallationExternalIssueBaseEndpoint as ExternalIssueBaseEndpoint,
 )
-from sentry.sentry_apps.api.bases.sentryapps import (
-    rpc_user_from_request,
-)
 from sentry.sentry_apps.services.cell import sentry_app_cell_service
 from sentry.sentry_apps.utils.errors import SentryAppError
+from sentry.users.services.user.serial import serialize_generic_user
 
 
 @control_silo_endpoint
@@ -28,11 +26,15 @@ class SentryAppInstallationExternalIssueDetailsEndpoint(ExternalIssueBaseEndpoin
                 status_code=400,
             )
 
+        rpc_user = serialize_generic_user(request.user)
+        if rpc_user is None:
+            return Response({"detail": "Authentication credentials were not provided."}, status=401)
+
         result = sentry_app_cell_service.delete_external_issue(
             organization_id=installation.organization_id,
             installation=installation,
             external_issue_id=external_issue_id,
-            user=rpc_user_from_request(request),
+            user=rpc_user,
         )
 
         if result.error:
